refactor(work2): extract resolveTaskType to shared module
Move duplicate resolveTaskType logic from mark-task-progress.js and implement-gate.js into resolve-task-type.js for DRY principle.
